Added conditional around part of paratest.server.pl that concatenates memory leak logs so that it only does so if there are memory leak logs.
git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15889 3a8e244f-b0f2-452b-bcba-4c88e055c3ca

+++ test/paratest.server.pl
@@ -257,8 +257,10 @@ sub feed_nodes {
     }
 
     $endtime = `date`; chomp $endtime;
-    systemd("cat $logdir/tmp.*.memleaks > $memleaks");
-    systemd("rm -f $logdir/tmp.*.memleaks");
+    if ($memleaksflag) {
+        systemd("cat $logdir/tmp.*.memleaks > $memleaks");
+        systemd("rm -f $logdir/tmp.*.memleaks");
+    }
     collect_logs ($fin_logfile, @logs);
 }
